Migrated ci from Travis CI to GitHub Actions workflow.
This repository is currently using https://travis-ci.org/. However, https://travis-ci.org/ was terminated on June 15, 2021.

> Since June 15th, 2021, the building on travis-ci.org is ceased. Please use travis-ci.com from now on.
> https://travis-ci.org/github/mikker/passwordless

Therefore, I moved to GitHub Action to restart CI.

The ruby 2.5 was excluded from CI because it is EOL.

I tried to set the repository secret `CODECOV_TOKEN` so that reports will be sent to codecov,
but I can't set the token, so I checked test_helper so that CI can be run even if it is not specified.