Fix claiming tokens readme
The `Supporting UUID primary keys` section was added to the middle of
the `Claiming tokens` section.

It made the last half the of the `Claiming tokens` section (`Upgrading
an existing Rails app`) slighlty confusing, since it was out-of-place.

Alternativley, we could also remove the `Upgrading an existing Rails
app` sub-section.  It looks like the `Claiming tokens` feature was added
few years ago.