Allow re-sending of magic link emails
When the config setting Passwordless.allow_token_resend = true a visit
to /users/sign_in/TOKEN_URL for an expired or already claimed user will
return a page with a button which can be used to resend an email with
this token. Extra care is taken to record a query parameter
`destination_path` into the current session, so opening the new email
immediately from the same browser redirects you to the proper location.