Add input token page to the sign in flow (#159)
Since building Passwordless I've changed my opinion slightly on what the best Passwordless sign in flow is.

Most of the time, I find it preferable to, instead of clicking a magic link, copy-paste a token directly. I prefer it because:

It lets me stay in the current browser tab
It's easier if I'm on a computer but I check my email on my phone
More than that, having it this way doesn't require removing the old way.

So that's where Passwordless is going.