Add default flash notice on SessionsController#destroy
This commit adds a default flash notice to destroy action
while also keeping the flexibility to customize redirect options
via config.

When creating a session in case of success there is a flash notice
displayed while when destroying a session there was not flash noticed
displayed in case of success.

Also having a successful flash message for destroying a session makes
the system tests in the apps using this gem a bit easier to check
if the logout happened successfully.